Revert "Botservices"
